Renamed main window file.
